First pass at structuring sanity's ruby gem
Had started this back when we were doing the POC to test out the API
functionality and did a little more yesterday.

This doesn't do a whole lot other than provide some basic structure for
the gem's API interactions. No tests & minimal docs. Mutating a document
with the class methods works so far.

We don't have to go this direction, but we can.