Remove activesupport dependency
This commit removes activesupport and replaces the methods needed with
our own hand rolled methods. Some taken from ActiveSupport and some much
less robust.

Using refinements here can ensure that we're not ruining the global
namespace and not causing any potential conflicts with ActiveSupport
should this be used in a Rails based application